Simplifed theme use by removing inherite widget
By extending ThemeData, we can avoid the use
of an inherited widget on top of the tree.
It simplifies the user inteaction with the
package since less is needed to use it and
the theming flow is consistent with the normal
one.
