+++ lib/src/widgets/gtk_popup_menu.dart
@@ -40,7 +40,7 @@ class _GtkPopupMenuState extends State<GtkPopupMenu> {
           context: context,
           barrierColor: Colors.transparent,
           contentDyOffset: 4,
-          backgroundColor: GnomeTheme.of(context).menus,
+          backgroundColor: Theme.of(context).menus,
           transitionDuration: const Duration(milliseconds: 150),
           bodyBuilder: (context) => SizedBox(child: widget.body),
           direction: PopoverDirection.top,
